+++ internal/controller/operator/factory/vmcluster/vmcluster.go
@@ -1527,6 +1527,10 @@ func buildVMauthLBDeployment(cr *vmv1beta1.VMCluster) (*appsv1.Deployment, error
 	if err != nil {
 		return nil, fmt.Errorf("cannot patch containers: %w", err)
 	}
+	strategyType := appsv1.RollingUpdateDeploymentStrategyType
+	if cr.Spec.RequestsLoadBalancer.Spec.UpdateStrategy != nil {
+		strategyType = *cr.Spec.RequestsLoadBalancer.Spec.UpdateStrategy
+	}
 	lbDep := &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace:       cr.Namespace,
@@ -1538,6 +1542,10 @@ func buildVMauthLBDeployment(cr *vmv1beta1.VMCluster) (*appsv1.Deployment, error
 		Spec: appsv1.DeploymentSpec{
 			Selector: &metav1.LabelSelector{
 				MatchLabels: cr.VMAuthLBSelectorLabels(),
+			},
+			Strategy: appsv1.DeploymentStrategy{
+				Type:          strategyType,
+				RollingUpdate: cr.Spec.RequestsLoadBalancer.Spec.RollingUpdate,
 			},
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{