+++ internal/controller/operator/factory/vtcluster/vmauth_lb.go
@@ -179,6 +179,11 @@ func buildVMauthLBDeployment(cr *vmv1.VTCluster) (*appsv1.Deployment, error) {
 	if err != nil {
 		return nil, fmt.Errorf("cannot patch containers: %w", err)
 	}
+	strategyType := appsv1.RollingUpdateDeploymentStrategyType
+	if cr.Spec.RequestsLoadBalancer.Spec.UpdateStrategy != nil {
+		strategyType = *cr.Spec.RequestsLoadBalancer.Spec.UpdateStrategy
+	}
+
 	lbDep := &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace:       cr.Namespace,
@@ -190,6 +195,10 @@ func buildVMauthLBDeployment(cr *vmv1.VTCluster) (*appsv1.Deployment, error) {
 		Spec: appsv1.DeploymentSpec{
 			Selector: &metav1.LabelSelector{
 				MatchLabels: cr.VMAuthLBSelectorLabels(),
+			},
+			Strategy: appsv1.DeploymentStrategy{
+				Type:          strategyType,
+				RollingUpdate: cr.Spec.RequestsLoadBalancer.Spec.RollingUpdate,
 			},
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{