+++ docs/CHANGELOG.md
@@ -20,6 +20,7 @@ aliases:
 
 * FEATURE: [vmanomaly](https://docs.victoriametrics.com/anomaly-detection/): support ui preset mode, support `vlogs` reader type. See [#1532](https://github.com/VictoriaMetrics/operator/issues/1538).
 * FEATURE: [vmoperator](https://docs.victoriametrics.com/operator/): add `unhealthyPodEvictionPolicy` to the `podDisruptionBudget` specification. See this issue [#1534](https://github.com/VictoriaMetrics/operator/issues/1534) for details.
+* FEATURE: [vmoperator](https://docs.victoriametrics.com/operator/): add `rollingUpdate` and `updateStrategy` fields to `VMAuth.spec`, `VMCluster.spec.requestsLoadBalancer.spec`, `VLCluster.spec.requestsLoadBalancer.spec` and `VTCluster.spec.requestsLoadBalancer.spec`. See this issue [#1540](https://github.com/VictoriaMetrics/operator/issues/1540) for details.
 * FEATURE: [vmoperator](https://docs.victoriametrics.com/operator/): preserve 3rd party `labels` on object during `reconcile`. Previously, operator allowed to keep only `managedMetadata.labels`. See this issue [#1533](https://github.com/VictoriaMetrics/operator/issues/1533). Thanks to the @lllamnyp
 
 * BUGFIX: [vmoperator](https://docs.victoriametrics.com/operator/): fix an issue where the return value from a couple of controllers was always `nil`. See [#1532](https://github.com/VictoriaMetrics/operator/pull/1532) for details.